change: move the script to execute.js
