+++ index.js
@@ -1,30 +1,6 @@
 const axios = require('axios');
 const FormData = require('form-data');
 const codec = require('string-codec');
-const readlineSync = require("readline-sync");
-
-let letCookiesTalenta = "_identity=identityCookie"
-cookiesTalenta = "PHPSESSID=asdasdasdasdasdasd"
-let latitude = "-6.861928521923624"
-let longitude = "107.63578698474662"
-
-const dataInput = () => {
-  const inLat = readlineSync.question(`Set Latitude (default: ${latitude}): `)
-  const inLong = readlineSync.question(`Set Longitude (default: ${longitude}): `)
-  const inCookie = readlineSync.question(`Set Latitude (default: ${cookiesTalenta}): `)
-
-  if(inLat!=""){
-    latitude = inLat;
-  }
-
-  if(inLong!=""){
-    longitude = inLong;
-  }
-
-  if(inCookie!=""){
-    cookiesTalenta = inCookie;
-  }
-}
 
 const prepForm = (obj) => {
   const { long, lat, desc, cookies, isCheckOut = false } = obj;
@@ -53,7 +29,6 @@ const prepForm = (obj) => {
 };
 
 const attendancePost = async (obj) => {
-  const inputPrompt = dataInput();
   const config = prepForm(obj);
   const resp = await axios(config);
 
@@ -71,11 +46,4 @@ const clockOut = async (obj) => {
 module.exports = {
   clockIn,
   clockOut
-}
-
-
-if(process.argv[2] == "clockin"){
-  console.log(clockIn({ lat: latitude, long: longitude, cookies: cookiesTalenta, desc: 'Talenta Clock In' }));
-}else if(process.argv[2] == "clockout"){
-  console.log(clockOut({ lat: latitude, long: longitude, cookies: cookiesTalenta, desc: 'Talenta Clock In' }));
 }
