+++ package.json
@@ -4,8 +4,8 @@
   "description": "API HR Talenta for ClockIn and ClockOut",
   "main": "index.js",
   "scripts": {
-    "clockin": "node index.js clockin",
-    "clockout": "node index.js clockout",
+    "clockin": "node execute.js clockin",
+    "clockout": "node execute.js clockout",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -20,7 +20,6 @@
   "homepage": "https://github.com/ans-4175/talenta-api#readme",
   "dependencies": {
     "axios": "^0.22.0",
-    "cheerio": "^1.0.0-rc.10",
     "form-data": "^4.0.0",
     "readline-sync": "^1.4.10",
     "string-codec": "^0.1.0"
